requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
pytest==8.0.2
pytest-cov==4.1.0
gunicorn==21.2.0
redis==5.0.1
Flask-Caching==2.1.0 